+++ module/types/params_test.go
@@ -25,13 +25,13 @@ func (suite *ParamsTestSuite) SetupTest() {
 		Limit:          sdk.NewInt(10000000000000),
 		TimeLimited:    false,
 		TimeBasedLimit: sdk.ZeroInt(),
-		TimePeriod:     time.Hour,
+		TimePeriod:     int64(time.Hour),
 	}
 	supply2 := types.SupplyLimit{
 		Limit:          sdk.NewInt(10000000000000),
 		TimeLimited:    true,
 		TimeBasedLimit: sdk.NewInt(100000000000),
-		TimePeriod:     time.Hour * 24,
+		TimePeriod:     int64(time.Hour * 24),
 	}
 	suite.supply = append(suite.supply, supply1, supply2)
 	return
@@ -189,7 +189,7 @@ func (suite *ParamsTestSuite) TestParamValidation() {
 			args: args{
 				assetParams: types.AssetParams{types.NewAssetParam(
 					"bnb", 714,
-					types.SupplyLimit{sdk.NewInt(-10000000000000), false, time.Hour, sdk.ZeroInt()}, true,
+					types.SupplyLimit{sdk.NewInt(-10000000000000), false, int64(time.Hour), sdk.ZeroInt()}, true,
 					suite.addr, sdk.NewInt(1000), sdk.NewInt(100000000), sdk.NewInt(100000000000),
 					types.DefaultSwapBlockTimestamp, types.DefaultSwapTimeSpan)},
 			},
@@ -201,7 +201,7 @@ func (suite *ParamsTestSuite) TestParamValidation() {
 			args: args{
 				assetParams: types.AssetParams{types.NewAssetParam(
 					"bnb", 714,
-					types.SupplyLimit{sdk.NewInt(10000000000000), false, time.Hour, sdk.NewInt(-10000000000000)}, true,
+					types.SupplyLimit{sdk.NewInt(10000000000000), false, int64(time.Hour), sdk.NewInt(-10000000000000)}, true,
 					suite.addr, sdk.NewInt(1000), sdk.NewInt(100000000), sdk.NewInt(100000000000),
 					types.DefaultSwapBlockTimestamp, types.DefaultSwapTimeSpan)},
 			},
@@ -213,7 +213,7 @@ func (suite *ParamsTestSuite) TestParamValidation() {
 			args: args{
 				assetParams: types.AssetParams{types.NewAssetParam(
 					"bnb", 714,
-					types.SupplyLimit{sdk.NewInt(10000000000000), true, time.Hour, sdk.NewInt(100000000000000)},
+					types.SupplyLimit{sdk.NewInt(10000000000000), true, int64(time.Hour), sdk.NewInt(100000000000000)},
 					true,
 					suite.addr, sdk.NewInt(1000), sdk.NewInt(100000000), sdk.NewInt(100000000000),
 					types.DefaultSwapBlockTimestamp, types.DefaultSwapTimeSpan)},
